Split api and core source files
To clean up the code a bit, split out the API functions called by
NCCL and the core functions called by the two transports.
