+++ src/compat_v2.c
@@ -4,33 +4,7 @@
 
 #include "config.h"
 
-#include "nccl_ofi.h"
-
-ncclResult_t nccl_net_ofi_pciPath_v2(int dev, char** path)
-{
-	ncclNetProperties_t props_latest;
-	ncclResult_t ret = ncclSuccess;
-
-	ret = nccl_net_ofi_getProperties(dev, &props_latest);
-
-	if (ret == ncclSuccess)
-		*path = props_latest.pciPath;
-
-	return ret;
-}
-
-ncclResult_t nccl_net_ofi_ptrSupport_v2(int dev, int *supportedTypes)
-{
-	ncclNetProperties_t props_latest;
-	ncclResult_t ret = ncclSuccess;
-
-	ret = nccl_net_ofi_getProperties(dev, &props_latest);
-
-	if (ret == ncclSuccess)
-		*supportedTypes = props_latest.ptrSupport;
-
-	return ret;
-}
+#include "nccl_ofi_api.h"
 
 const ncclNet_v2_t ncclNetPlugin_v2 = {
 	.name = "AWS Libfabric",
